refactor calendar elements into nested hierarchy
