+++ src/components/pebble-base-calendar.ts
@@ -3,12 +3,7 @@ import { property } from "lit/decorators.js";
 import {
   isSameDay,
   isWithinInterval,
-  Day,
   format,
-  startOfWeek,
-  startOfMonth,
-  addDays,
-  eachDayOfInterval,
 } from "date-fns";
 import { CalendarEvent } from "../utils/calendar-utils";
 import { COLOR_CSS_VARS } from "../utils/colors";
@@ -18,14 +13,6 @@ import { HomeAssistant } from "../types";
 
 export abstract class PebbleBaseCalendar extends LitElement {
   @property({ attribute: false }) public hass!: HomeAssistant;
-
-  @property({ attribute: false }) protected weekStartsOn: Day;
-
-  @property({ attribute: false }) protected numWeeks?: number;
-
-  @property({ attribute: false }) protected monthCalendarStart?: "current_week" | "start_of_month";
-
-  @property({ attribute: false }) protected weekCalendarStart?: "current_week" | "current_day";
 
   @property({ attribute: false }) protected events: CalendarEvent[];
 
@@ -37,12 +24,8 @@ export abstract class PebbleBaseCalendar extends LitElement {
 
   @property({ attribute: false }) protected localize: (key: LocalizationKey) => string;
 
-
   constructor() {
     super();
-    this.weekStartsOn = 0;
-    this.numWeeks = 12;
-    this.monthCalendarStart = "current_week";
     this.events = [];
     this.localize = (arg) => arg;
   }
@@ -64,34 +47,6 @@ export abstract class PebbleBaseCalendar extends LitElement {
           (b.allDay ? 0 : b.start.getHours() * 60 + b.start.getMinutes()),
       );
   }
-
-  protected generateWeeksInMonth() {
-    const numWeeks = this.numWeeks ?? 12;
-    const today = Date.now();
-    const startPosition = this.monthCalendarStart ?? "current_week";
-    const weekStartsOn = +(this.weekStartsOn ?? 0) as Day;
-
-    let startDate: Date;
-    if (startPosition === "start_of_month") {
-      startDate = startOfMonth(today);
-    } else {
-      // current_week - start from the beginning of the current week
-      startDate = startOfWeek(today, { weekStartsOn });
-    }
-
-    const firstWeekStart = startOfWeek(startDate, { weekStartsOn });
-
-    // Generate all weeks in the continuous range
-    const weeks = [];
-    for (let i = 0; i < numWeeks; i++) {
-      const weekStart = addDays(firstWeekStart, i * 7);
-      const weekEnd = addDays(weekStart, 6);
-      weeks.push(eachDayOfInterval({ start: weekStart, end: weekEnd }));
-    }
-
-    return weeks;
-  }
-
 
   protected renderForecast(forecast?: ForecastAttribute) {
     if (!forecast) {
@@ -165,14 +120,12 @@ export abstract class PebbleBaseCalendar extends LitElement {
                 : nothing
             }
           </div>
-
-        
         </div>
       </ha-dialog>
     `;
   }
 
-  static get sharedStyles(): CSSResultGroup {
+  static get baseStyles(): CSSResultGroup {
     return [
       COLOR_CSS_VARS,
       css`
@@ -187,101 +140,6 @@ export abstract class PebbleBaseCalendar extends LitElement {
           height: 100%;
           display: grid;
           align-items: end;
-        }
-
-        .calendar {
-          display: grid;
-          grid-template-columns: repeat(7, minmax(0, 1fr));
-          grid-auto-rows: min-content;
-          height: 100%;
-          padding: 0 12px;
-          overflow: visible;
-        }
-
-        .calendar-container {
-          height: 100%;
-          overflow: hidden;
-          position: relative;
-        }
-
-        .calendar-scroll-area {
-          height: min(100%, calc(100vh - var(--header-height)));
-          overflow-y: scroll;
-          overflow-x: hidden;
-          scroll-behavior: smooth;
-          scrollbar-width: thin;
-          scroll-snap-type: y mandatory;
-        }
-
-        .calendar-scroll-area::-webkit-scrollbar {
-          width: 6px;
-        }
-
-        .calendar-scroll-area::-webkit-scrollbar-track {
-          background: rgba(0, 0, 0, 0.1);
-          border-radius: 3px;
-        }
-
-        .calendar-scroll-area::-webkit-scrollbar-thumb {
-          background: rgba(0, 0, 0, 0.3);
-          border-radius: 3px;
-        }
-
-        .calendar-scroll-area::-webkit-scrollbar-thumb:hover {
-          background: rgba(0, 0, 0, 0.5);
-        }
-
-        .calendar-header {
-          display: grid;
-          grid-template-columns: repeat(7, minmax(0, 1fr));
-          position: sticky;
-          top: 0;
-          z-index: 10;
-          border-bottom: 2px solid var(--divider-color, #e0e0e0);
-          padding: 0 12px;
-        }
-
-        .day {
-          min-height: 100px;
-          position: relative;
-          overflow: visible;
-        }
-
-        .day-name,
-        .day {
-          margin: var(--day-margin, 5px);
-          padding: 0 0 12px 0;
-          border-bottom: 2px solid #ccc;
-          font-size: 2em;
-          line-height: 120%;
-          text-align: right;
-        }
-
-        .day-name {
-          display: inline-block;
-          padding-right: 10px;
-        }
-
-        .day-name.active-day {
-          border-bottom: 3px solid var(--primary-text-color);
-          font-weight: bold;
-        }
-
-        .month {
-          display: inline-block;
-        }
-
-        .numeral {
-          display: inline-block;
-          padding: 10px;
-          border-radius: 100px;
-          min-width: 1.25em;
-          text-align: center;
-        }
-
-        .numeral.today {
-          background-color: rgb(68, 68, 68);
-          padding: 10px;
         }
 
         .past {
